@@ -37,13 +37,13 @@
   },
   "license": "Apache-2.0",
   "devDependencies": {
-    "@types/jest": "^24.0.15",
-    "@types/node": "12.6.2",
-    "jest": "^24.8.0",
-    "jsii": "^0.14.0",
-    "jsii-pacmak": "^0.14.0",
-    "json-schema-to-typescript": "^6.1.3",
-    "typescript": "^3.5.3"
+    "@types/jest": "^24.0.18",
+    "@types/node": "12.7.11",
+    "jest": "^24.9.0",
+    "jsii": "^0.18.0",
+    "jsii-pacmak": "^0.18.0",
+    "json-schema-to-typescript": "^7.1.0",
+    "typescript": "^3.6.3"
   },
   "jest": {
     "moduleFileExtensions": [
@@ -54,10 +54,11 @@
     "yaml"
   ],
   "dependencies": {
-    "@aws-cdk/core": "^1.0.0",
-    "yaml": "^1.6.0"
+    "@aws-cdk/core": "^1.11.0",
+    "js-yaml": "^3.13.1",
+    "yaml": "^1.7.0"
   },
   "peerDependencies": {
-    "@aws-cdk/core": "^1.0.0"
+    "@aws-cdk/core": "^1.11.0"
   }
 }
