@@ -1,10 +1,11 @@
 import { Construct, ISynthesisSession } from '@aws-cdk/core';
 import fs = require('fs');
 import path = require('path');
-import { KubResource } from './resource';
-const YAML = require('yaml');
+import { ApiObject } from './api-object';
+const yaml = require('js-yaml');
 
-export class Stack extends Construct {
+
+export class Chart extends Construct {
 
   /**
    * The name of the stack's YAML file as emitted into the cloud assembly
@@ -21,11 +22,11 @@ export class Stack extends Construct {
   protected synthesize(session: ISynthesisSession) {
     const resources = new Array<any>();
 
-    for (const resource of this.node.findAll().filter(x => x instanceof KubResource)) {
-      resources.push((resource as KubResource).render());
+    for (const resource of this.node.findAll().filter(x => x instanceof ApiObject)) {
+      resources.push((resource as ApiObject).render());
     }
 
-    const doc = resources.map(r => YAML.stringify(r)).join('---\n');
+    const doc = resources.map(r => yaml.dump(r, { skipInvalid: true })).join('---\n');
     fs.writeFileSync(path.join(session.assembly.outdir, this.assemblyFileName), doc);
   }
 }
