@@ -1,4 +1,4 @@
 export * from './service';
 export * from './deployment';
-export * from './resource';
-export * from './stack';
+export * from './api-object';
+export * from './chart';
