@@ -1,4 +1,4 @@
-import { Stack, KubResource } from '../lib';
+import { Chart, ApiObject } from '../lib';
 import { App, Construct } from '@aws-cdk/core';
 import { synth } from './util';
 
@@ -7,7 +7,7 @@ test('empty stack', () => {
   const app = new App();
 
   // WHEN
-  const stack = new Stack(app, 'empty');
+  const stack = new Chart(app, 'empty');
 
   // THEN
   expect(synth(stack)).toMatchSnapshot();
@@ -16,17 +16,17 @@ test('empty stack', () => {
 test('output includes all synthesized resources', () => {
   // GIVEN
   const app = new App();
-  const stack = new Stack(app, 'test');
+  const stack = new Chart(app, 'test');
 
   // WHEN
-  new KubResource(stack, 'Resource1', { kind: 'Resource1', apiVersion: 'v1' });
-  new KubResource(stack, 'Resource2', { kind: 'Resource2', apiVersion: 'v1' });
-  new KubResource(stack, 'Resource3', { kind: 'Resource3', apiVersion: 'v1' });
+  new ApiObject(stack, 'Resource1', { kind: 'Resource1', apiVersion: 'v1' });
+  new ApiObject(stack, 'Resource2', { kind: 'Resource2', apiVersion: 'v1' });
+  new ApiObject(stack, 'Resource3', { kind: 'Resource3', apiVersion: 'v1' });
 
   // also, subscopes
   const scope = new Construct(stack, 'scope');
-  new KubResource(scope, 'Resource1', { kind: 'Resource1', apiVersion: 'v1' });
-  new KubResource(scope, 'Resource2', { kind: 'Resource2', apiVersion: 'v1' });
+  new ApiObject(scope, 'Resource1', { kind: 'Resource1', apiVersion: 'v1' });
+  new ApiObject(scope, 'Resource2', { kind: 'Resource2', apiVersion: 'v1' });
 
   // THEN
   expect(synth(stack)).toMatchSnapshot();
